patchcommander rework (#69)
When sampling with more complex patch configurations, patches along the border of the image were not always sampled correctly. For example, with center=True, patches with the center outside the image were not sampled, ignoring the patches from the bottom row of the image. Additionally, with overlap, additional patches around the borders of the images still contained tissue. For example  in the case of a config with half overlapping slides, the patch half a patch size to the upper left of the (0, 0) coordinate was not sampled , while the bottom quarter of this patch still contained the image. These issues are fixed in this PR. For potential required lookup for the visual validation plots for this PR, check intermediate commit 7910ed83d74f6e1aec3d0c19fce990d9c0309310/